Async cashing of tab-completion
In order to make tab-completion faster this starts a background-worker
everytime the program is run to update a local cache file.

Due to Kube-library not beeing QUITE thread-safe or at least not
multiprocessing-thread safe I use os.fork and some quite questionable
code-flow to make it work... Ehh tell me if it breaks impressivley for
you!